Prevent jetpants clone_slave on existing slave targets: Prevent clone_slave from failing in a confusing way in the following edge case scenario -- 1. A MySQL slave instance has crashed. MySQL is not running, and the node's    data has been cleared out. The node needs to be re-cloned but will remain    in the same pool. 2. An asset tracker plugin is in use, and the plugin automatically imputes    the master/slaves of a non-running MySQL instance by checking the asset    tracker pool data at probe time. 3. The asset tracker still lists the crashed node as being in its appropriate    pool, despite the node having crashed.
In this case, attempting to use clone_slave will fail on the call to
DB#resume_replication on the target. The subtle reason is that #2 causes
DB#change_master_to to be a no-op, since the node's @master will already have
been set to the right value already.

This small fix prevents the situation by disallowing existing slaves from
being targets of a clone_slave operation.