Smarter :like context when claiming spare slaves in some cases: In "jetpants shard_split" and "jetpants shard_cutover", when we claim spares to make slaves for the new shard(s), supply a :like param of the new shard's master.
This way, a custom plugin could implemented special logic to assist with a
MySQL upgrade if desired -- ie, it could make :role => :master always use
a new version box (regardless of :like), so that any newly-created shards
will be created using the newer version of MySQL from the start. Combined
with the change in this commit, in this situation the shard will now also
correctly get new-version slaves matching the master.